refactor add package interview and divise packages
