+++ src/main/java/ma/nttdata/externals/module/interview/entity/Interview.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import ma.nttdata.externals.module.candidate.entity.Candidate;
+import ma.nttdata.externals.module.offer.entity.Offer;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
