+++ src/main/java/ma/nttdata/externals/module/offer/repository/OfferRepository.java
@@ -1,5 +1,5 @@
-package ma.nttdata.externals.module.interview.repository;
-import ma.nttdata.externals.module.interview.entity.Offer;
+package ma.nttdata.externals.module.offer.repository;
+import ma.nttdata.externals.module.offer.entity.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
