+++ src/main/resources/application-dev.yml
@@ -20,7 +20,7 @@ spring:
   jpa:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
     hibernate:
-      ddl-auto: none  # Let Flyway handle all schema creation
+      ddl-auto: validate   # Use "update" for development, but "validate" for production
     show-sql: true
     properties:
       hibernate:
@@ -30,23 +30,13 @@ spring:
             non_contextual_creation: true
 
   flyway:
-    enabled: false
+    enabled: true
     locations: classpath:db/migration
     baseline-on-migrate: true
 
   docker:
     compose:
       enabled: false
-
-  sql:
-    init:
-      mode: always
-      schema-locations:
-        - classpath:db/migration/V1__Create_user_and_candidate_tables.sql
-        - classpath:db/migration/V2__Offers__and__interviews__tables.sql # Point to the migrtaion file
-      data-locations:
-        - classpath:db/insertion/Data.sql # Point to the insertion of data file
-      continue-on-error: false
 
 logging:
   level:
@@ -62,4 +52,3 @@ app:
   ai:
     server:
       url: http://localhost:8080
-
